+++ apps/admin-dashboard/pages/chief/index.vue
@@ -6,6 +6,8 @@
 
 <script setup>
 definePageMeta({
-  layout: false
+  layout: false,
+  middleware: 'chief'
 })
+
 </script>
